fix(memory) Introduce `Memory.buffer()` to generate new view.
The previous memory API was broken by design. To read the memory data,
we need to generate a new view every time, or to at worst refresh the
inner byte buffer. We decide to adopt the first approach, and
introduce a new `Memory.buffer` method that returns a fresh direct
`ByteBuffer`. Thus, `Memory` no longer has a `read` and `write`
methods.